Move dump resources arg parsing into a separate file
